Created BowlingGame class which is going to manage, let comunicate the bowlingball class and bowlingframe class. Created test for the BowlingFrame class to test the method saveScore, which save a given score in its scores array.
Incapsulated bowlingball and bowlingframe class inside the BowlingGame class. It contain a frames array which will keep all the frame of the game, and a turn variable which will keep count of the turnation.
sketched the rollAndSave method in the BowlingGame class, which will take care of roll the ball and save the score in the correct frame.

All test green.